+++ src/library/components/table/index.tsx
@@ -23,6 +23,8 @@ import { update } from '../../utils/immutable';
 import { QueryString } from '@borvik/querystring';
 import { DeepPartial } from '@borvik/use-querystate/dist/types';
 import { TableFooter } from './footer';
+import { TableContextProvider, useTableSelector } from './contexts';
+import isEqual from 'lodash/isEqual';
 
 const preMDR_RenderWarned: Record<string, boolean> = {};
 const preMDR_WidthWarned: Record<string, boolean> = {};
@@ -30,7 +32,7 @@ const primaryKeyWarned: {[x:string]: boolean} = {};
 const fixedLeftWarned: Record<string, boolean> = {};
 const fixedRightWarned: Record<string, boolean> = {};
 
-export const DataTable = function DataTable<T, FooterData extends T = T>({paginate = 'both', quickEditPosition = 'both', hideSearchForm = false, defaultFilter, methodRef, passColumnsToQuery, ...props}: PropsWithChildren<DataTableProperties<T, FooterData>>) {
+const DataTableCore = function DataTableCore<T, FooterData extends T = T>({paginate = 'both', quickEditPosition = 'both', hideSearchForm = false, defaultFilter, methodRef, passColumnsToQuery, ...props}: PropsWithChildren<DataTableProperties<T, FooterData>>) {
   const canGroupBy = !!props.canGroupBy && !!props.multiColumnSorts;
 
   /**
@@ -186,9 +188,6 @@ export const DataTable = function DataTable<T, FooterData extends T = T>({pagina
   const canEdit = typeof props.onSaveQuickEdit === 'function' && columnData.hasEditor && (columnData.primaryKeyCount === 1 || typeof props.getRowKey === 'function');
   const canSelectRows = !!props.canSelectRows && (columnData.primaryKeyCount === 1 || typeof props.getRowKey === 'function');
 
-  const [editFormData, setFormData] = useState<EditFormData>({});
-  const [selectedRows, setSelectedRows] = useState<Record<string | number, T>>({});
-
   const [pagination, setPagination] = useParsedQs<Pagination, DeepPartial<Pagination>>(
     {page: 1, perPage: props.paginateOptions?.defaultPerPage ?? 10},
     (qsPagination) => ({
@@ -272,8 +271,7 @@ export const DataTable = function DataTable<T, FooterData extends T = T>({pagina
     }
   );
 
-  const [isEditing, setEditing] = useState(false);
-  const [isSavingQuickEdit, setSaving] = useState(false);
+  const [, setCtxData] = useTableSelector(() => ({}), isEqual);
   const [editCount, setEditCount] = useState(0);
 
   const [stateDataList, setDataList] = useState<DataFnResult<T[], FooterData[]>>({ data: [], total: 0 });
@@ -413,7 +411,7 @@ export const DataTable = function DataTable<T, FooterData extends T = T>({pagina
   const onSaveQuickEdit = useCallback(async (data: QuickEditFormData<T>) => {
     try {
       if (!!propOnSave && Object.keys(data).length) {
-        setSaving(true)
+        setCtxData({ isSavingQuickEdit: true });
         let rowsToSave = Object.keys(data);
         let primaryColumn = columnData.actualColumns.find(c => c.isPrimaryKey)!;
         let originalData: QuickEditFormData<T> = {};
@@ -427,16 +425,16 @@ export const DataTable = function DataTable<T, FooterData extends T = T>({pagina
         }
 
         await propOnSave(data, originalData);
-        setFormData({});
+        setCtxData({ editData: {} });
         setEditCount(c => c + 1);
       }
-      setEditing(false);
+      setCtxData({ isEditing: false });
     }
     catch {
       // avoid unhandled exception
     }
     finally {
-      setSaving(false);
+      setCtxData({ isSavingQuickEdit: false });
     }
   }, [propOnSave, currentData, actualColumnsComparator, propGetRowKey]);
 
@@ -447,7 +445,7 @@ export const DataTable = function DataTable<T, FooterData extends T = T>({pagina
       let values = Object.values(selection);
       props.onSelectionChange(keys, values);
     }
-    setSelectedRows(selection);
+    setCtxData({ selectedRows: selection });
   }
 
   function setAllSelected(selectAll: boolean) {
@@ -467,17 +465,33 @@ export const DataTable = function DataTable<T, FooterData extends T = T>({pagina
 
   function setRowSelected(row: any, rowIndex: number) {
     let rowKey = getRowKey(row, rowIndex, columnData.actualColumns, props.getRowKey);
-    let exists = typeof selectedRows[rowKey] !== 'undefined';
+    setCtxData(prev => {
+      const exists = typeof prev.selectedRows[rowKey] !== 'undefined';
+      let newState = prev;
+      if (!exists) {
+        newState = update(prev, {
+          selectedRows: {
+            [rowKey]: { $set: row },
+          }
+        });
+      } else {
+        newState = update(prev, {
+          selectedRows: {
+            $unset: [rowKey]
+          }
+        });
+      }
+
+      if (typeof props.onSelectionChange === 'function') {
+        // raise
+        let keys = Object.keys(newState.selectedRows);
+        let values = Object.values(newState.selectedRows);
+        props.onSelectionChange(keys, values);
+      }
+      return newState;
+    });
     
-    if (!exists) {
-      doSetSelectedRows(update(selectedRows, {
-        [rowKey]: { $set: row },
-      }));
-    } else {
-      doSetSelectedRows(update(selectedRows, {
-        $unset: [rowKey]
-      }));
-    }
+    
   }
 
   const searchFormOnSearch = useCallback((query: string) => {
@@ -569,13 +583,7 @@ export const DataTable = function DataTable<T, FooterData extends T = T>({pagina
       multiColumnSorts: props.multiColumnSorts ?? false,
       filter,
       filterSettings: props.filterSettings,
-      isEditing,
-      isSavingQuickEdit,
-      editData: editFormData,
-      editMode: props.editMode ?? 'default',
       canSelectRows,
-      selectedRows,
-      setFormData: setFormData,
       setFilter,
       setColumnSort,
       setAllSelected,
@@ -617,7 +625,6 @@ export const DataTable = function DataTable<T, FooterData extends T = T>({pagina
                 quickEditPosition={quickEditPosition}
                 buttons={{
                   quickEdit: <TableEditorButton
-                    setEditing={setEditing}
                     canEdit={canEdit}
                   />,
                   filter: <FilterButton />,
@@ -667,7 +674,6 @@ export const DataTable = function DataTable<T, FooterData extends T = T>({pagina
             quickEditPosition={quickEditPosition}
             buttons={{
               quickEdit: <TableEditorButton
-                setEditing={setEditing}
                 canEdit={canEdit}
               />,
               filter: <FilterButton />,
@@ -699,3 +705,11 @@ export const DataTable = function DataTable<T, FooterData extends T = T>({pagina
  *    <section>
  *      <pages>
  */
+
+export const DataTable = function DataTable<T, FooterData extends T = T>({editMode, ...props}: PropsWithChildren<DataTableProperties<T, FooterData>>) {
+  return (
+    <TableContextProvider editMode={editMode ?? 'default'}>
+      <DataTableCore {...props} />
+    </TableContextProvider>
+  );
+};